api/requirements.txt
tensorflow==2.6.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.6.0
matplotlib
numpy
